Bug: 688805 Reviewer: Self Comment: Fix bug in getActivity method.

+++ trunk/ssf/main/src/org/kablink/teaming/module/report/impl/ReportModuleImpl.java
@@ -529,7 +529,7 @@ public class ReportModuleImpl extends HibernateDaoSupport implements ReportModul
 					}
 				}
 				if (entity != null && !entity.isDeleted()) 
-					list.add(new ActivityInfo(entity, (Integer) col[3], (Date) col[4]));
+					list.add(new ActivityInfo(entity, ((Long) col[3]).intValue(), (Date) col[4]));
 			} catch (Exception ignoreAccess) {continue;}
 			if (list.size() >= returnCount.intValue()) break;
 		}
